Refactor: Extract vehicle and building logic to hooks
Move vehicle and building helper functions into their own hooks for better organization and reusability.
