+++ src/components/game/buildingHelpers.ts
@@ -1,0 +1,115 @@
+import { useCallback, useMemo } from 'react';
+import { Tile, BuildingType } from '@/types/game';
+import { getBuildingSize } from '@/lib/simulation';
+
+export function useBuildingHelpers(grid: Tile[][], gridSize: number) {
+  const parkBuildingsSet = useMemo(() => new Set<BuildingType>([
+    'park_large', 'baseball_field_small', 'football_field',
+    'mini_golf_course', 'go_kart_track', 'amphitheater', 'greenhouse_garden',
+    'marina_docks_small', 'roller_coaster_small', 'mountain_lodge', 'playground_large', 'mountain_trailhead'
+  ]), []);
+
+  const isPartOfMultiTileBuilding = useCallback((gridX: number, gridY: number): boolean => {
+    const maxSize = 4;
+    
+    for (let dy = 0; dy < maxSize; dy++) {
+      for (let dx = 0; dx < maxSize; dx++) {
+        const originX = gridX - dx;
+        const originY = gridY - dy;
+        
+        if (originX >= 0 && originX < gridSize && originY >= 0 && originY < gridSize) {
+          const originTile = grid[originY][originX];
+          const buildingSize = getBuildingSize(originTile.building.type);
+          
+          if (buildingSize.width > 1 || buildingSize.height > 1) {
+            if (gridX >= originX && gridX < originX + buildingSize.width &&
+                gridY >= originY && gridY < originY + buildingSize.height) {
+              return true;
+            }
+          }
+        }
+      }
+    }
+    
+    return false;
+  }, [grid, gridSize]);
+
+  const findBuildingOrigin = useCallback((gridX: number, gridY: number): { originX: number; originY: number; buildingType: BuildingType } | null => {
+    const maxSize = 4;
+    
+    const tile = grid[gridY]?.[gridX];
+    if (!tile) return null;
+    
+    if (tile.building.type !== 'empty' && 
+        tile.building.type !== 'grass' && 
+        tile.building.type !== 'water' && 
+        tile.building.type !== 'road' && 
+        tile.building.type !== 'tree') {
+      const size = getBuildingSize(tile.building.type);
+      if (size.width > 1 || size.height > 1) {
+        return { originX: gridX, originY: gridY, buildingType: tile.building.type };
+      }
+      return null;
+    }
+    
+    if (tile.building.type === 'empty') {
+      for (let dy = 0; dy < maxSize; dy++) {
+        for (let dx = 0; dx < maxSize; dx++) {
+          const originX = gridX - dx;
+          const originY = gridY - dy;
+          
+          if (originX >= 0 && originX < gridSize && originY >= 0 && originY < gridSize) {
+            const originTile = grid[originY][originX];
+            
+            if (originTile.building.type !== 'empty' && 
+                originTile.building.type !== 'grass' &&
+                originTile.building.type !== 'water' &&
+                originTile.building.type !== 'road' &&
+                originTile.building.type !== 'tree') {
+              const size = getBuildingSize(originTile.building.type);
+              
+              if (size.width > 1 || size.height > 1) {
+                if (gridX >= originX && gridX < originX + size.width &&
+                    gridY >= originY && gridY < originY + size.height) {
+                  return { originX, originY, buildingType: originTile.building.type };
+                }
+              }
+            }
+          }
+        }
+      }
+    }
+    
+    return null;
+  }, [grid, gridSize]);
+
+  const isPartOfParkBuilding = useCallback((gridX: number, gridY: number): boolean => {
+    const maxSize = 4;
+
+    for (let dy = 0; dy < maxSize; dy++) {
+      for (let dx = 0; dx < maxSize; dx++) {
+        const originX = gridX - dx;
+        const originY = gridY - dy;
+
+        if (originX >= 0 && originX < gridSize && originY >= 0 && originY < gridSize) {
+          const originTile = grid[originY][originX];
+
+          if (parkBuildingsSet.has(originTile.building.type)) {
+            const buildingSize = getBuildingSize(originTile.building.type);
+            if (gridX >= originX && gridX < originX + buildingSize.width &&
+                gridY >= originY && gridY < originY + buildingSize.height) {
+              return true;
+            }
+          }
+        }
+      }
+    }
+    return false;
+  }, [grid, gridSize, parkBuildingsSet]);
+
+  return {
+    isPartOfMultiTileBuilding,
+    findBuildingOrigin,
+    isPartOfParkBuilding,
+  };
+}
